+++ templates/master/proxy-lex/index.js
@@ -18,6 +18,30 @@ module.exports={
         }]
       }
     },
+    "LexStatusLambda": {
+      "Type": "AWS::Lambda::Function",
+      "Properties": {
+        "Code": {
+            "ZipFile":fs.readFileSync(__dirname+'/status.js','utf8')
+        },
+        "Environment": {
+            "Variables":{
+                STATUS_KEY:"status.json",
+                STATUS_BUCKET:{"Ref":"BuildStatusBucket"},
+                BOT_NAME:{"Ref":"LexBot"},
+            }
+        },
+        "Handler": "index.handler",
+        "MemorySize": "128",
+        "Role": {"Fn::GetAtt": ["LexProxyLambdaRole","Arn"]},
+        "Runtime": "nodejs6.10",
+        "Timeout": 300,
+        "Tags":[{
+            Key:"Type",
+            Value:"Api"
+        }]
+      }
+    },
     "LexProxyLambdaRole": {
       "Type": "AWS::IAM::Role",
       "Properties": {
@@ -37,7 +61,20 @@ module.exports={
         "ManagedPolicyArns": [
           "arn:aws:iam::aws:policy/service-role/AWSLambdaBasicExecutionRole",
           "arn:aws:iam::aws:policy/AmazonLexFullAccess"
-        ]
+        ],
+        "Policies":[{
+            "PolicyName":"Access",
+            "PolicyDocument": {
+              "Version": "2012-10-17",
+              "Statement": [{
+                    "Effect": "Allow",
+                    "Action": [
+                        "s3:Get*"
+                    ],
+                    "Resource":[{"Fn::Sub":"arn:aws:s3:::${BuildStatusBucket}*"}]
+              }]
+            }
+        }]
       }
     }
 }